- Movendo métodos para a ViewModel - Clean Code.
