+++ prj/app/src/main/java/com/brenobatista/desafiozup/Models/ServiceData.kt
@@ -3,7 +3,6 @@ package com.brenobatista.desafiozup.Models
 import com.google.gson.annotations.SerializedName
 
 data class ServiceData(
-
     @SerializedName("serviceName")
     val serviceName: String
 )
