+++ prj/app/src/main/java/com/brenobatista/desafiozup/Services/IWebService.kt
@@ -1,9 +1,21 @@
 package com.brenobatista.desafiozup.Services
 
+import com.brenobatista.desafiozup.Models.ServicesResult
+import com.brenobatista.desafiozup.Models.UsersResult
 import retrofit2.Call
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.http.GET
+
+interface IWebService {
+    @GET("/zupbank/api/servicetypes")
+    fun getServices(
+    ): Call<ServicesResult>
+
+    @GET("/zupbank/api/appusers")
+    fun getUsers(
+    ): Call<UsersResult>
+}
 
 fun createService(): IWebService {
     val retrofit = Retrofit.Builder()
@@ -12,14 +24,4 @@ fun createService(): IWebService {
         .build()
 
     return retrofit.create(IWebService::class.java)
-}
-
-interface IWebService {
-    @GET("/zupbank/api/servicetypes")
-    fun searchServices(
-    ): Call<ServicesResult>
-
-    @GET("/zupbank/api/appusers")
-    fun searchUsers(
-    ): Call<UsersResult>
 }
