+++ prj/app/src/main/java/com/brenobatista/desafiozup/ViewModels/LoginFragment.kt
@@ -9,11 +9,6 @@ import android.widget.Button
 import androidx.navigation.fragment.findNavController
 import com.brenobatista.desafiozup.R
 
-/**
- * A simple [Fragment] subclass.
- * Use the [LoginFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class LoginFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -24,7 +19,6 @@ class LoginFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_login, container, false)
     }
