+++ prj/app/src/main/java/com/brenobatista/desafiozup/Views/MainListViewActivity.kt
@@ -1,24 +1,19 @@
 package com.brenobatista.desafiozup.Views
 
 import android.os.Bundle
-import android.util.Log
 import android.view.Menu
 import android.widget.SearchView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.brenobatista.desafiozup.Data.ReposAdapter
 import com.brenobatista.desafiozup.R
-import com.brenobatista.desafiozup.Services.ServicesResult
-import com.brenobatista.desafiozup.Services.createService
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
+import com.brenobatista.desafiozup.ViewModels.MainListViewModel
 
 class MainListViewActivity : AppCompatActivity() {
 
-    private val reposAdapter = ReposAdapter()
-    private var filterAdapter = ReposAdapter()
+    companion object{
+        private val vm = MainListViewModel()
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -27,31 +22,10 @@ class MainListViewActivity : AppCompatActivity() {
         // Cria a RecyclerView e seta seu adapter
         val list: RecyclerView = findViewById(R.id.recyclerViewServicos)
         list.layoutManager = LinearLayoutManager(this)
-        list.adapter = reposAdapter
+        list.adapter = vm.reposAdapter
 
         // Criação dos serviços de consumo de APIs
-        val service = createService()
-        service.searchServices().enqueue(object : Callback<ServicesResult> {
-            override fun onFailure(call: Call<ServicesResult>, t: Throwable) {
-                // Tratar exceções
-                Log.wtf("GET_DATA_ERROR", "Erro ao obter dados do serviço", t)
-            }
-
-            override fun onResponse(
-                call: Call<ServicesResult>,
-                response: Response<ServicesResult>
-            ) {
-                try {
-                    val data = response.body()
-                    val repos = data?.services
-
-                    if (repos?.count()!! > 0)
-                        reposAdapter.submitList(repos)
-
-                } catch (e: Exception) {
-                }
-            }
-        })
+        vm.initSearchService()
     }
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
@@ -70,13 +44,13 @@ class MainListViewActivity : AppCompatActivity() {
 
                     val searchPattern = searchQuery.toString().toLowerCase()
 
-                    var lst = reposAdapter.currentList.filter { p ->
+                    var lst = vm.reposAdapter.currentList.filter { p ->
                         p.serviceName.toLowerCase().contains(searchPattern)
                     }
-                    filterAdapter.submitList(lst)
+                    vm.filterAdapter.submitList(lst)
 
                     val view: RecyclerView = findViewById(R.id.recyclerViewServicos)
-                    view.adapter = filterAdapter
+                    view.adapter = vm.filterAdapter
 
                     return true
                 }
@@ -84,9 +58,5 @@ class MainListViewActivity : AppCompatActivity() {
         }
 
         return super.onCreateOptionsMenu(menu)
-    }
-
-    private fun filter() {
-
     }
 }
